added_launch, fixed subtask
